chore: box-shadow and refactoring
