+++ src/components/theme/theme.css
@@ -1,7 +1,7 @@
 .color-theme-toggle {
   position: absolute;
-  bottom: 5px;
-  right: 5px;
+  bottom: 10px;
+  right: 10px;
 }
 
 .color-theme-toggle--selected {
@@ -18,11 +18,18 @@
 
 .color-theme-toggle__button {
   border-radius: 50%;
-  border: 4px solid white;
-  width: 50px;
-  height: 50px;
+  width: 45px;
+  height: 45px;
   text-decoration: none;
   margin: 5px;
+  border: 0;
+}
+
+.color-theme-toggle__button--selected {
+  border: 5px solid white;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+  width: 50px;
+  height: 50px;
 }
 
 .color-theme-toggle__button--pink {
